verified comments
