+++ src/components/Header.js
@@ -1,27 +1,16 @@
-import { useState } from 'react' //pulling useState from the react library so we can use it directly
+import React from "react"; //brings in the react functionality to this file
 
-const AddTodo = ({ onAdd }) => { //create the AddTodo component and pass in the onAdd function from the prop by destructing
-    const [item, setItem] = useState('') //declaring the item variable and setItem function for the useState and setting items default to an empty string
+const Header = (props) => { //create header component function passing in the props object
+  return ( //return the structure we created for the component
+    <div> {/** open a div */}
+      <h1 style={headerStyle}>{props.title}</h1> {/** pass in the headerStyle object which gives the component(h2) custom styles and pass in the title from the prop */}
+    </div> //closes div
+  );
+};
 
-    const onSubmit = (e) => { //a function literal that takes in the event response
-        e.preventDefault() // prevents/stops the default behaviours of the event
-        onAdd({item}) //invoking the onAdd funtion and passing it the item variable (as an object?) which will add it to our todo list since it is the addTodo function declared in the app.js file
+const headerStyle = { //create a constant variable called headerStyle
+  fontSize: "40px", //set the font size to 40px
+  textDecoration: "underline", //set the text decoration to underline
+};
 
-        setItem('') //resets the item variable to an empty string so it is ready to recieve a new item.
-    }
-
-    return ( //return the component structure we created
-        <form onSubmit={onSubmit}> {/**inserts a form and passes our onSubmit function to it */}
-            <label>Add Todo:</label>
-            <input 
-                type="text" 
-                placeholder="Todo Item"
-                value={item}  //{/** pass the name variable from the useState to it */}
-                onChange={ (e) => setItem(e.target.value)} //creates an anonymous function the receives the event response and uses setItem to update item to the value written given to it by the event response.
-            /> 
-            <input type="submit" value="Add Todo" />
-        </form>
-    )
-}
-
-export default AddTodo //exports the created component/app as the default item which means we can simply import it into to another file by just referencing the file.
+export default Header; //exports the created component/app as the default item which means we can simply import it into to another file by just referencing the file.
