+++ src/components/Todo.js
@@ -1,5 +1,5 @@
 
-const Todo = ({ todo }) => { //creates a constant variable that contains a function that destructures the passed in prop grabing the passed in todo item/object
+const Todo = ({ todo }) => { //creates a component by declaring a constant variable that contains a function that destructures the passed in prop grabing the passed in todo item/object
     return ( //returns the component we are defining inbetween the parenthesis
         <div> {/* place open div element tag*/}
             <h2>{todo.item}</h2> {/** place h2 element containing the list item passed to it */}
